pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incident_env"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
server = "incident_env.server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["incident_env*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
